+++ spec/spec_helper.rb
@@ -10,3 +10,15 @@ if ENV['CI']
 end
 require 'onlyoffice_digitalocean_wrapper'
 require 'securerandom'
+
+# Any existing Droplet name
+# @return [String] name of droplet
+def existing_droplet
+  'wrata.teamlab.info'
+end
+
+# Any existing Droplet IP
+# @return [String] ip of droplet
+def existing_droplet_ip
+  '138.197.104.24'
+end